Time: 7 ms (8.22%), Space: 22.9 MB (10.55%) - LeetHub

+++ 0026-remove-duplicates-from-sorted-array/0026-remove-duplicates-from-sorted-array.cpp
@@ -26,7 +26,7 @@ class Solution {
 public:
     int removeDuplicates(vector<int>& nums) {
         short sizeNums = nums.size();
-        int *indcies=new int[sizeNums];
+        short *indcies=new short[sizeNums];
         fill_n(indcies, sizeNums, -1);
         short start = 0;
         for (short i = sizeNums - 1; i > 0; i--)
